Mention the 'autofire_button' key binding action. (#135)
Some games assign different meanings (e.g. different weapons) to short
and long presses of the same fire button. For such a game it is
helpful to be able to generate both a long press of Fire and many
rapid-fire short presses, without having to pause the emulator and
toggle autofire on and off constantly.

Eventually I found out that I could achieve this by binding an extra
key (I used Right Shift, keeping Right Control for long presses of
fire) to the useful value "action_joy_1_autofire_button". But that
value wasn't mentioned in the docs anywhere at all, and I had to read
the source code to find it. It's very useful, so I think it _should_
be documented!